Favorite item refacto : split Favorite item and booking button
